All done ver0.4

+++ Final/esb/static/esb/create.js
@@ -3,6 +3,76 @@ var uploadedImages = [];
 var num = 0;
 var count = 1;
 
+document.addEventListener("DOMContentLoaded", () => {
+    // Show first image of product
+    if (document.querySelector("#index1")) {
+        document.querySelector("#index1").style.display = "block";
+        const btnDelete = document.querySelector("#index-delete1");
+        if (btnDelete) {
+            btnDelete.style.display = "block";
+        }
+    }
+    const images = document.querySelectorAll(".img-thumbnail");
+
+    // Prev disabled initially.
+    const btnPrev = document.querySelector("#prev-image");
+    if (btnPrev) {
+        if (count <= 1) {
+            btnPrev.disabled = true;
+        } else {
+            btnPrev.disabled = false;
+        }
+    }
+    // Next disabled initially if 1 image only.
+    const btnNext = document.querySelector("#next-image");
+    if (btnNext) {
+        if (count >= images.length) {
+            btnNext.disabled = true;
+        } else {
+            btnNext.disabled = false;
+        }
+    }
+    // Initial model
+    const inputTitle = document.getElementsByName("title")[0];
+    if (inputTitle) {
+        if (inputTitle.value === "title") {
+            inputTitle.value = "";
+        } else {
+            document.querySelector(".btn.btn-primary").innerHTML = "Edit"; // Change the button name 'Create' to 'Edit'
+        }
+    }
+
+    const inputPrice = document.getElementsByName("price")[0];
+    if (inputPrice) {
+        if (inputPrice.value === "0.00") {
+            inputPrice.value = "";
+        }
+    }
+
+    const inputDescription = document.getElementsByName("description")[0];
+    if (inputDescription) {
+        if (inputDescription.value === "description") {
+            inputDescription.value = "";
+        }
+    }
+
+    // Auto Select Category when Edit
+    const typeSelect = document.getElementById("type-category");
+    const categoryType = document.getElementById("hidden-category-type");
+    const categoryTitle = document.getElementById("hidden-category-title");
+    if (categoryType) {
+        const defaultOption = typeSelect.querySelector(
+            `option[value='${categoryType.value}']`
+        );
+        if (defaultOption) {
+            defaultOption.selected = true;
+        }
+    }
+    if (categoryTitle) {
+        updateSubCategories();
+    }
+});
+
 function uploadImage() {
     const fileInput = document.getElementById("image-input");
     const file = fileInput.files[0];
@@ -20,6 +90,9 @@ function uploadImage() {
             uploadedImages.push(imageURL);
             product_id.value = data.product_id; // Hidden input value provide Product ID
             displayImages(data.image_id);
+
+            // Make fileinput become null, To prevent the first image upload again when cancel the second uploading.
+            fileInput.value = "";
         })
         .catch((error) => console.error(error));
 }
@@ -44,12 +117,14 @@ function displayImages(image_id) {
     num += 1;
     image.className = "img-thumbnail";
     btnDelete.className = "btn btn-danger";
-    image.id = "index" + numImages.toString();
+
+    image.id = `index${numImages.toString()}`;
     btnDelete.id = "index-delete" + numImages.toString();
     btnDelete.innerHTML = "Delete";
     btnDelete.style.position = "absolute";
     btnDelete.onclick = function () {
-        deleteImage(image_id);
+        const id_num = image.id.split("x")[1];
+        deleteImage(image_id, id_num);
     };
     // Show img uploaded
     image.style.display = "block";
@@ -57,9 +132,13 @@ function displayImages(image_id) {
 
     // Set position of count
     count = numImages;
-    // Uploaded image at last position, so next should be disabled, and prev be able.
+    // Uploaded image at last position, so next should be disabled, and prev be able(prev be disabled if first image).
     document.querySelector("#next-image").disabled = true;
-    document.querySelector("#prev-image").disabled = false;
+    if (count <= 1) {
+        document.querySelector("#prev-image").disabled = true;
+    } else {
+        document.querySelector("#prev-image").disabled = false;
+    }
     // Add uplaoded image to Img View Area
     previewDiv.appendChild(image);
     previewDiv.appendChild(btnDelete);
@@ -84,10 +163,12 @@ function toggleImage(action) {
         count = count - 1;
         image = document.querySelector(`#index${count}`);
         btn = document.querySelector(`#index-delete${count}`);
+        console.log(count);
     } else if (action == "next") {
         count = count + 1;
         image = document.querySelector(`#index${count}`);
         btn = document.querySelector(`#index-delete${count}`);
+        console.log(count);
     }
     image.style.display = "block";
     btn.style.display = "block";
@@ -109,7 +190,14 @@ function toggleImage(action) {
     }
 }
 
-function deleteImage(image_id) {
+function deleteImage(image_id, idNum) {
+    // Confirm
+    const confirmed = confirm("Are you sure you want to delete this image?");
+
+    if (!confirmed) {
+        return;
+    }
+
     // Delete image through its ID
     fetch("/delete_image", {
         method: "DELETE",
@@ -120,6 +208,58 @@ function deleteImage(image_id) {
         .then((response) => response.json())
         .then((result) => {
             console.log(result);
+            // Length of Images before remove image
+            const imagesLength =
+                document.querySelectorAll(".img-thumbnail").length;
+            // After deleted database's data, remove the display too.
+            const imagePrepareToRemove = document.getElementById(
+                `index${idNum}`
+            );
+            const buttonPrepareToRemove = document.getElementById(
+                `index-delete${idNum}`
+            );
+
+            imagePrepareToRemove.remove();
+            buttonPrepareToRemove.remove();
+
+            console.log("idNum:", idNum);
+            console.log("imagesLength:" + imagesLength);
+            const images = document.querySelectorAll(".img-thumbnail");
+            console.log("images.length:" + images.length);
+            if (imagesLength > idNum) {
+                console.log("images ok");
+                images.forEach((img) => {
+                    console.log(" each ok");
+                    const idImg = img.id.split("x")[1]; // Get the index number. e.g. Get 1 from index1, 2 from index2
+                    if (idImg > idNum) {
+                        const newID = "index" + (idImg - 1);
+                        const btnDelete = document.getElementById(
+                            `index-delete${idImg}`
+                        );
+
+                        img.id = newID;
+                        btnDelete.id = "index-delete" + (idImg - 1);
+                        console.log("count" + count);
+                    }
+                });
+            }
+            // Show first image if exists.
+            if (document.getElementById("index1")) {
+                document.getElementById("index1").style.display = "block";
+                document.getElementById("index-delete1").style.display =
+                    "block";
+            }
+            // First image, Prev become disabled.
+            const btnPrev = document.querySelector("#prev-image");
+            btnPrev.disabled = true;
+            // Next Image become disabled if there's only 1 image.
+            const btnNext = document.querySelector("#next-image");
+            count = 1;
+            if (count >= images.length) {
+                btnNext.disabled = true;
+            } else {
+                btnNext.disabled = false;
+            }
         });
 }
 
@@ -199,73 +339,3 @@ function updateProduct() {
         })
         .catch((error) => console.error(error));
 }
-
-document.addEventListener("DOMContentLoaded", () => {
-    // Show first image of product
-    if (document.querySelector("#index1")) {
-        document.querySelector("#index1").style.display = "block";
-        const btnDelete = document.querySelector("#index-delete1");
-        if (btnDelete) {
-            btnDelete.style.display = "block";
-        }
-    }
-    const images = document.querySelectorAll(".img-thumbnail");
-
-    // Prev disabled initially.
-    const btnPrev = document.querySelector("#prev-image");
-    if (btnPrev) {
-        if (count <= 1) {
-            btnPrev.disabled = true;
-        } else {
-            btnPrev.disabled = false;
-        }
-    }
-    // Next disabled initially if 1 image only.
-    const btnNext = document.querySelector("#next-image");
-    if (btnNext) {
-        if (count >= images.length) {
-            btnNext.disabled = true;
-        } else {
-            btnNext.disabled = false;
-        }
-    }
-    // Initial model
-    const inputTitle = document.getElementsByName("title")[0];
-    if (inputTitle) {
-        if (inputTitle.value === "title") {
-            inputTitle.value = "";
-        } else {
-            document.querySelector(".btn.btn-primary").innerHTML = "Edit"; // Change the button name 'Create' to 'Edit'
-        }
-    }
-
-    const inputPrice = document.getElementsByName("price")[0];
-    if (inputPrice) {
-        if (inputPrice.value === "0.00") {
-            inputPrice.value = "";
-        }
-    }
-
-    const inputDescription = document.getElementsByName("description")[0];
-    if (inputDescription) {
-        if (inputDescription.value === "description") {
-            inputDescription.value = "";
-        }
-    }
-
-    // Auto Select Category when Edit
-    const typeSelect = document.getElementById("type-category");
-    const categoryType = document.getElementById("hidden-category-type");
-    const categoryTitle = document.getElementById("hidden-category-title");
-    if (categoryType) {
-        const defaultOption = typeSelect.querySelector(
-            `option[value='${categoryType.value}']`
-        );
-        if (defaultOption) {
-            defaultOption.selected = true;
-        }
-    }
-    if (categoryTitle) {
-        updateSubCategories();
-    }
-});
